% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarc.R
\name{add_miss_qtr}
\alias{add_miss_qtr}
\title{Add missing quarter}
\usage{
add_miss_qtr(x)
}
\arguments{
\item{x}{a data.table with a column QY with data in format YYYYQX}
}
\value{
a vector pf data format YYYYQX with missing quarters filled in
}
\description{
Add missing quarter
}
