% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarc.R
\name{add_qy}
\alias{add_qy}
\title{Create a feature for Policy period in quarters}
\usage{
add_qy(dt, year_col, incept_date, add_miss)
}
\arguments{
\item{dt}{a data.table}

\item{year_col}{a field specif}

\item{incept_date}{a date field}

\item{add_miss}{optional. TRUE evaluates missing quarters see add_miss_qtr()}
}
\value{
a datatable with an additional field called QY with period in YYYYQX format
}
\description{
Create a feature for Policy period in quarters
}
