% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calc_lev}
\alias{calc_lev}
\title{Calculate Limited Expected Value}
\usage{
calc_lev(
  dt,
  limit,
  att_pt,
  k = 1,
  mu = 15.2830643838331,
  sigma = 1.74824394618441
)
}
\arguments{
\item{dt}{a datatable}

\item{limit}{a field with limit values}

\item{att_pt}{a field with attachment values}

\item{k}{moment of lognormal distribution}

\item{mu}{mean of log transform}

\item{sigma}{sd of log transform}
}
\value{
adds a column to the datatable with LEV
}
\description{
Calculate Limited Expected Value
}
