% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{clean_RLCC}
\alias{clean_RLCC}
\title{Convert NA to Not Categorized for Risk Factor Levels}
\usage{
clean_RLCC(x)
}
\arguments{
\item{x}{a character vector of Risk Factor Levels}
}
\value{

}
\description{
Convert NA to Not Categorized for Risk Factor Levels
}
