% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{convert_data_type}
\alias{convert_data_type}
\title{Transform Wide View data into correct data types for R}
\usage{
convert_data_type(dt, num_cols, date_cols, factor_cols)
}
\arguments{
\item{dt}{a data.table}

\item{num_cols}{character vector of list of identified numeric fields}

\item{date_cols}{character vector of list of identified date fields}

\item{factor_cols}{character vector of list of identified factor fields}
}
\value{

}
\description{
Transform Wide View data into correct data types for R
}
