% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{enrich_levels}
\alias{enrich_levels}
\title{Backfill missing categorical levels for Insured}
\usage{
enrich_levels(dt, ins_name, field)
}
\arguments{
\item{dt}{a data.table with Insured Name, and at least one risk factor}

\item{ins_name}{insured name field}

\item{field}{a categorical risk factor}
}
\value{
a datatable populated with unique insured names and the most frequently populated level from the data
}
\description{
Backfill missing categorical levels for Insured
}
