% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{inspect_results}
\alias{inspect_results}
\title{Compare Standardised name to original}
\usage{
inspect_results(org, ref)
}
\arguments{
\item{org}{original values}

\item{ref}{refined values}
}
\value{
datatable with 3 columns,original, edited and TRUE/FALSE to indicate a  match
}
\description{
Compare Standardised name to original
}
