% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{lev_log_norm}
\alias{lev_log_norm}
\title{Limited Expected Value of Lognormal}
\usage{
lev_log_norm(x, k, mu, sigma)
}
\arguments{
\item{x}{limit point}

\item{k}{moment (e.g. k=2 produces second moment)}

\item{mu}{mean of log transform}

\item{sigma}{std of log transform}
}
\value{
kth moment of lognormal limited at x
}
\description{
Limited Expected Value of Lognormal
}
\examples{
lev_log_norm(1000, 1,  15.2830643838331, 1.74824394618441 )

}
