% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{lev_log_norm_sdev}
\alias{lev_log_norm_sdev}
\title{Standard deviation of Limited Lognormal}
\usage{
lev_log_norm_sdev(p, mu, sigma)
}
\arguments{
\item{sigma}{}
}
\value{

}
\description{
Standard deviation of Limited Lognormal
}
