% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{name_preprocess}
\alias{name_preprocess}
\title{preprocess names}
\usage{
name_preprocess(x, clean_patt)
}
\arguments{
\item{x}{character string}
}
\value{
character string
}
\description{
preprocess names
}
