% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{quarter_year}
\alias{quarter_year}
\title{Quarter Year}
\usage{
quarter_year(x)

quarter_year(x)
}
\arguments{
\item{x}{month in digits (1-12)}
}
\value{
a character either Q1, Q2, Q3, or Q4

a character either Q1, Q2, Q3, or Q4
}
\description{
Quarter Year

Quarter Year
}
