% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarc.R
\name{RARC_Component}
\alias{RARC_Component}
\title{Clean up rate change components}
\usage{
RARC_Component(x)
}
\arguments{
\item{x}{a character vector of names}
}
\value{
a character vector
}
\description{
Clean up rate change components
}
