% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{refine_string}
\alias{refine_string}
\title{cluster names by similarity}
\usage{
refine_string(
  x,
  edit_threshold = 1,
  numgram = 2,
  ignore_strings = NULL,
  dict = NULL
)
}
\arguments{
\item{x}{Character vector, items to be potentially clustered and merged.}

\item{edit_threshold}{Numeric value, indicating the threshold at which a merge is performed, based on the sum of the edit values derived from param weight. Default value is 1. If this parameter is set to 0 or NA, then no approximate string matching will be done, and all merging will be based on strings that have identical ngram fingerprints.}

\item{numgram}{Numeric value, indicating the number of characters that will occupy each ngram token. Default value is 2.}

\item{ignore_strings}{Character vector, these strings will be ignored during the merging of values within vect. Default value is NULL.}

\item{dict}{Character vector, meant to act as a dictionary during the merging process. If any items within vect have a match in dict, then those items will always be edited to be identical to their match in dict. Default value is NULL.}
}
\value{
standardised names
}
\description{
cluster names by similarity
}
\examples{
refine_string(c("Acme Pizza, Inc.", "AcMe PiZzA, Inc.", "ACME PIZZA COMPANY", "acme pizza LLC"))
}
