% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{select_level}
\alias{select_level}
\title{Helper function to enrich levels to classify categorical levels}
\usage{
select_level(l, mfl)
}
\arguments{
\item{l}{a string. the level of a risk factor for an insured}

\item{mfl}{a string, the mostly frequently used  risk factor level for that insured}
}
\value{
a string.
}
\description{
Helper function to enrich levels to classify categorical levels
}
