% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{standardise_names}
\alias{standardise_names}
\title{Standardise names}
\usage{
standardise_names(dt, name_col, clean_patt)
}
\arguments{
\item{dt}{data.table}

\item{name_col}{vector of names to standardise}

\item{clean_patt}{string of patterns to remove before applying clustering algorithm}
}
\value{
data.table with an additional column called std_Insured_Name
}
\description{
Standardise names
}
