% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarc.R
\name{tidy_rarc}
\alias{tidy_rarc}
\title{Create a tidy RARC dataset}
\usage{
tidy_rarc(dt, premium_type, col_key)
}
\arguments{
\item{dt}{a data.table object}

\item{premium_type}{a character string of a premium term in the data}

\item{col_key}{a vector key categorical features e.g. name, cedant, coverage, industry}

\item{description}{this function calculates RARC using a calculated LEV to adjust for layer changes from bordereau data. It returns a long formatted data set
segmented by the features described in col_key and by period in the format YYYYQX.
The outputs are RARC, LEVChange (relies on pre-calculation of LEV using appropriate exposure curves)}
}
\value{
a tidy data set with RARC, LEV Change, Premium Change and RARC Weights according to the OG methodology
}
\description{
Create a tidy RARC dataset
}
